fix: resolve SyntaxErrors and schema inconsistencies
- Remove orphaned string literal in iOSEnterpriseAnalyzer.js (line 66)
- Rename conflicting variable in iOSEnterpriseChecks.js (functions -> extractedFunctions)
- Fix ai_gate schema: change 'rule' to 'rule_id' for consistency
- Update violations-detailed test to validate internal consistency only

All tests passing (45/45)
