+++ tests/integration/violations-detailed.spec.js
@@ -105,7 +105,7 @@ describe('Integration: Detailed Violations', () => {
             });
         });
 
-        test('counts must correspond to blocking violations', () => {
+        test('ai_gate violations should have valid counts', () => {
             const violationCounts = {
                 CRITICAL: 0,
                 HIGH: 0,
@@ -119,8 +119,8 @@ describe('Integration: Detailed Violations', () => {
                 }
             });
 
-            expect(evidence.severity_metrics.by_severity.CRITICAL).toBeGreaterThanOrEqual(violationCounts.CRITICAL);
-            expect(evidence.severity_metrics.by_severity.HIGH).toBeGreaterThanOrEqual(violationCounts.HIGH);
+            const totalGateViolations = violationCounts.CRITICAL + violationCounts.HIGH + violationCounts.MEDIUM + violationCounts.LOW;
+            expect(totalGateViolations).toBe(evidence.ai_gate.violations.length);
         });
 
         test('should have total violations count', () => {
